requirements.txt
keras==2.12.0
matplotlib==3.7.1
noise==1.2.2
numpy==1.23.5
Pillow~=9.5.0
rasterio~=1.3.8
tensorflow~=2.12.0